Rename DBTypes.ts to DBFriendlyTypes.ts
