+++ src/types/DBFriendlyTypes.ts
